Moved 'tags' field from Ingredient class to Module.
