+++ src/main/java/org/launchcode/sprinklespre/models/Module.java
@@ -6,6 +6,7 @@ import jakarta.persistence.ManyToOne;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Set;
 
 @Entity
 public class Module extends AbstractEntity{
@@ -19,6 +20,9 @@ public class Module extends AbstractEntity{
     @ManyToMany
     private List<Ingredient> ingredients = new ArrayList<>();
 
+    // What category is the main ingredient in it (vegetable, fruit, protein, etc)? Meant to make searching easier.
+    private Set<String> tags;
+
     boolean isCompleted;
 
     public Module() {}
@@ -30,5 +34,39 @@ public class Module extends AbstractEntity{
 
     void finishModule() {
         this.isCompleted = true;
+    }
+
+    //Getters and setters
+
+    public Course getCourse() {
+        return course;
+    }
+
+    public void setCourse(Course course) {
+        this.course = course;
+    }
+
+    public List<Tool> getTools() {
+        return tools;
+    }
+
+    public void setTools(List<Tool> tools) {
+        this.tools = tools;
+    }
+
+    public List<Ingredient> getIngredients() {
+        return ingredients;
+    }
+
+    public void setIngredients(List<Ingredient> ingredients) {
+        this.ingredients = ingredients;
+    }
+
+    public Set<String> getTags() {
+        return tags;
+    }
+
+    public void setTags(Set<String> tags) {
+        this.tags = tags;
     }
 }
